Why Validation issue
I used sublime Text 2 for this project I validated the html and css in http://validator.w3.org/ In the css, it said "Congratulations! No Error Found." and turned green. in the HTML, it said "This document was successfully checked as HTML5!" and turned green. Please explain the validation issue? Thanks, Hope